OSHMEM: Fixing coverity-related comments (prints, outputs, comments, etc)

+++ oshmem/shmem/c/shmem_alltoall.c
@@ -172,7 +172,7 @@ SHMEM_TYPE_ALLTOALLS(_alltoalls64, sizeof(uint64_t))
 #define SHMEM_TYPE_TEAM_ALLTOALL(type_name, type, code, postfix)    \
     int  shmem##type_name##postfix(shmem_team_t team, type *dest, const type *source, size_t nelems)   \
     {                                                               \
-        size_t rc = 0;                                              \
+        int rc = 0;                                                 \
                                                                     \
         RUNTIME_CHECK_INIT();                                       \
                                                                     \
@@ -212,10 +212,10 @@ SHMEM_TYPE_TEAM_ALLTOALL(_ptrdiff, ptrdiff_t, SHMEM_PTRDIFF_T, _alltoall)
 SHMEM_TYPE_TEAM_ALLTOALL(, void, SHMEM_BYTE, _alltoallmem)
 
 
-#define SHMEM_TYPE_TEAM_ALLTOALLS(type_name, type, code, postfix)    \
+#define SHMEM_TYPE_TEAM_ALLTOALLS(type_name, type, code, postfix)   \
     int  shmem##type_name##postfix(shmem_team_t team, type *dest, const type *source, ptrdiff_t dst, ptrdiff_t sst,  size_t nelems)   \
     {                                                               \
-        size_t rc = 0;                                              \
+        int rc = 0;                                                 \
                                                                     \
         RUNTIME_CHECK_INIT();                                       \
                                                                     \
